pyproject.toml
[project]
name = "ai-coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.49.0",
    "langchain>=1.0.5",
    "langchain-groq>=1.0.0",
    "langgraph>=1.0.2",
]
